Improve filerep logging
In many cases, when a segment sets a fault, or receives a transition message from another source, no breadcrumbs are left behind to suggest how or why this happened. Now, when a client connection is created to transition a segment from one state to another, we will log that connection had taken place. In addition, any time that a segment sets a filerep fault, we will print a stack trace to help us understand why it has done so.