WIP: Disable two-phase commit for vacuum
This commit is an attempt to liberate VACUUM from two phase
commit. During VACUUM, we do not actually write transaction ids to
anything, or make any other logical changes to tables, so it should be
safe to run VACUUM without a two-phase commit.

The merge with Postgres 9.0 has brought in the concept of 'relmapped'
tables. Tables like 'pg_class' are used to point to the relfilenodes
that contain any given relation, but this is problematic when trying
to swap out the relfile node of 'pg_class' itself.  The merge has also
brought in the use CLUSTER for VACUUM FULL. If you try to run VACUUM
FULL on 'pg_class', the relfile itself is swapped when it is vacuumed,
and the relmap is updated with a pointer to the new version of the
table. Postgres doesn't actually support PREPARE for changes to the
relmap, so instead of adding support for it in Greenplum, we have
opted to disable two-phase commit for VACUUM on heap tables.