Add dispatch to collation creation commands
Make CREATE COLLATION and pg_import_system_collations() parallel aware by
dispatching collation creation to the QEs.

In order for collations to work correctly, we need to be sure that every
collation that is created on the QD is also installed on the QEs, and that the
OID matches in every database. We will take advantage of two phase commit to
prevent collations from being created if there is a problem adding it on any
QE. In upstream, collations are created during initdb, but this won't work for
GPDB, because while initdb is running there is no way to be sure that every
segment has the same locales installed.

We disable collation creation during initdb, and make it the responsibility of
the system administrator to initialize any needed collations by either running
a CREATE COLLATION command, or running the pg_import_system_collations() UDF.

Co-authored-by: Jim Doty <jdoty@pivotal.io>