Avoid infinite autovacuum respawn loop
In GPDB we only run autovacuum on databases that do not allow connections,
like template0. When a database that *does* allow connections is older than
autovacuum_freeze_max_age, the autovacuum launcher starts a worker and
immediately exits.  The worker, finding a database that is older than
autovacuum_freeze_max_age, immediately starts the launcher so that the
next database can be processed. We only want to do this for databases
that do not allow connections, or we'll end up signaling for another
autovacuum worker to start that will not have any work to do, entering
an endless loop.

In order to prevent this from happening, we added a flag to
SetTransactionIdLimit() to prevent signaling an autovacuum worker start
when there will not be any work to do.

Co-authored-by: Adam Berlin <aberlin@pivotal.io>