Restrict folding of stable functions during planning
Stable functions need not be folded in case the query is simple enough
(e.g. pl/pgsql simple expression, or "select foo()").

Without this patch, every time a simple expression is evaluated, a new
plan and a new ExprState are generated. Because we only clean up the
expression states created for simple expressions at the end of
transaction, this will result in millions of expression states, one per
tuple, being generated that are not reused after they are initially
created.

This patch is taken from Heikki's review comment:
https://github.com/greenplum-db/gpdb/pull/5947#issuecomment-428103480

This patch also removes GPDB specific update for the test `guc.sql`.
Prior to this commit, since we were aggressively folding stable
functions, `current_setting()` was evaluated at plan time causing a diff
from upstream. This patch realigns the test with upstream.

Thanks to Jesse for valuable guidance and carving out the test case.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>
Co-authored-by: Heikki Linnakangas <hlinnakangas@pivotal.io>
Co-authored-by: Taylor Vesely <tvesely@pivotal.io>