Add an an integration test extension for the planner
Add a new extension that plans a query string and inspects the output of
the resulting PlannedStmt. This style test should help us define
expectations about what happens during planning from the perspective of
the executor or other components dependent on the planner. This approach
was inspired by `src/test/modules/pred_test` in upstream.

Co-authored-by: Adam Berlin <aberlin@pivotal.io>
Co-authored-by: Melanie Plageman <mplageman@pivotal.io>