Remove unnecessary JAVA_HOME sourcing
And do not source /opt/gcc_env.sh.

Now that we are using openjdk from the Centos 6 repositories in the
build image, setting the JAVA_HOME environment variable should not be
necessary. Additionally, the gcc environment is now expected to be
available without the need of sourcing /opt/gcc_env.sh.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>
Co-authored-by: Ben Christel <bchristel@pivotal.io>