Initialize cluster with primary/mirror replication slots.
This replication slot is used for WAL replication between primary and
mirror segments, and also master and standby.  The replication slot is
created when a mirror / standby segment is initialized using
pg_basebackup.  The replication slot is used by primary to keep track
of the WAL flush location reported by the mirror.  When the mirror
disconnects, it allows the primary to retain enough WAL so that the
mirror can catchup after reconnecting in future.

- defaults max_wal_senders to 10 to allow for basebackup to spin up
senders matching upstream
- defaults max_replication_slots to 10 instead of 0
- changes gp_basebackup to create a replication slot when a slot name is
provided during gpinitsystem
- changes gp_basebackup to use streaming replication during gpinitsystem
- creates and uses replication slot during full recovery

Note: We intend to reason more deeply with the default guc settings in
a later feature.

Co-authored-by: David Kimura <dkimura@pivotal.io>
Co-authored-by: Asim R P <apraveen@pivotal.io>