Update time zone data files to tzdata release 2018h
DST law changes in Kazakhstan and Alaska.
Historical corrections for Kwajalein, Nauru, Guam, Hong Kong, Chuuk,
Kosrae, Majuro, Palau, and Pohnpei.

This is just the tzdata.zi from the upstream IANA database. This should
be identical to what will eventually be added in upstream Postgres. Any
changes in the zic compiler in 2018h are not included, and will be
deferred until they are added in upstream Postgres.