Still dispatch DefineIndex even if this is part of alter table
The original motivation of putting off dispatching index creation and
dispatching (and executing) the entirety of an `ALTER TABLE` was to
enable `ALTER TABLE ... ALTER COLUMN ... SET DATA TYPE ...`. But there
are other (arguably sad) reasons we have to revert that change:

An `ALTER TABLE` statement on its own doesn't necessarily explicitly specify
all the indexes it creates. We cannot let QE pick a name for an index or we
risk that a QE arriving at a different conclusion about the name from that on
the QD.

Case in point: `ALTER TABLE ... ADD UNIQUE ...` on a partioned table may create
multiple indexes.