WIP: Adjust tablespace path handling
This commit links pg_tblspc to a <dbid> subdirectory of the tablespace, rather
than to the path of the tablespace directly, and removes the <dbid> from the
`tablespace_version_directory()`

Before this commit, the full path to an arbitrary file would look like:
  pg_tblspc/spcoid/ -> /<tablespace_location>/GPDB_MAJORVER_CATVER_db<dbid>/dboid/relfilenode
e.g.
  pg_tblspc/20981/ -> /data1/tsp1/GPDB_6_201902061_db1/19849/192814

After this commit, the full path to an abitrary file will look like this:
  pg_tblspc/spcoid/ -> /<tablespace_location>/<dbid>/GPDB_MAJORVER_CATVER/dboid/relfilenode
e.g.
  pg_tblspc/20981/ -> /data1/tsp1/1/GPDB_6_201902061/19849/192814

Motivation:

When tablespaces were aligned to upstream postgres, while removing
filespaces, we added the `tablespace_version_directory()` function to
supply each segment with a unique tablespace directory name. This was
accomplished by appending the 'magic' `GpIdentity.dbid` global variable
to the `GP_TABLESPACE_VERSION_DIRECTORY` in `tablespace_version_directory()`.

This is problematic for several reasons- but perhaps most severely is
the fact that in order to use any code in libpgcommon.so that references
this value, you need to first set the `GpIdentity.dbid` global,
otherwise any functions that deal with tablespaces will be broken in
unpredictable ways.

An example is pg_rewind- where `GetRelationPath()` will not return a valid
relation unless you repeatedly toggle the `GpIdentity.dbid` between the
value of the source or target segment dependant on the context of which
relfiles are being examined.