pg_basebackup: Use new tablespace layout
The server now removes the source dbid from the end of any tablespace symlink
target found in its response to BASE_BACKUP: This means that it removes it from
the tablespace header as well as any tar directory entries for a given
tablespace.  The pg_basebackup client now adds the target dbid at the end of
the symlink target returned from the server response in order to create the
correct symlink:
<target_datadir>/pg_tblspc/<tablespace_oid> -> <tablespace_location>/<target_db_id>

Note: All of the above is applicable to user-defined tablespaces.

Also, we renamed some of the ddl objects in the isolation2 tests for
tablespaces as they were too long. Also, reducing the length of the LOCATION
clause of the tablespace object helped us avoid the tar header limit of 100
characters for symlinks.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>
Co-authored-by: Taylor Vesely <tvesely@pivotal.io>
Co-authored-by: Adam Berlin <aberlin@pivotal.io>
(cherry picked from commit e81dd542edc0a5c39b68964d6bc6548fab4ffd45)