Modify pg_rewind to handle new tablespace layout.
* Add tests around pg_rewind's behavior with tablespaces.
* pg_rewind now lops off the source dbid from the symlink target before
populating the filemap.
* pg_rewind now appends the target dbid to the symlink target when
creating it and also creates the symlink target directory. Both of these
actions are performed as a separate action: `create_tablespace_layout`.
We distinguish tablespace symlinks with the `is_gp_tablespace` flag.
* pg_rewind now fetches the target dbid by invoking the postgres binary.
* Add test for pg_rewind with tablespaces which tests for objects
created after promotion of the replica. We introduced a new hook in the
pg_rewind suite: `before_master_restart_after_rewind`, which executes
right after pg_rewind completes and the master is restarted.
* Add test for pg_rewind with tablepace objects created before standby
promotion.
* Add pg_rewind test for objects removed after standby promotion

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>
Co-authored-by: Adam Berlin <aberlin@pivotal.io>