Do not get visibilitymap_count for AO relations
The relallvisible field in pg_class contains the number of heap pages
that are marked as all visible for any given heap relation. This value
is used by the planner to determine the cost of running an
IndexOnlyScan.

In theory, it is possible to use the visibility map of an AO table to
enable a index-only scan for AO tables. If there is no entry in the
pg_aovisimap table, then all tuples in the aoseg file up to the EOF will
be visible, and and index-only scan should not need to visit the AO
relation. It's not clear that enabling IndexOnlyScans will buy us
anything with AO relations, so it makes sense to just disable it for
now.

Until and unless we decide to enable index only scans of AO tables,
always set relallvisible to 0 for AO relations. This should prevent the
planner from choosing an index only scan for AO relations.

Like heap tables, AO tables have a visibility map, but the AO
visibilitymap serves a different purpose than the visibilitymap in heap
tables. In Heap tables, the visibility map is an optimization used to
determine which tuples can be returned directly from an index without
needing to consult with the heap table. Heap visibility maps do not
affect the visibility of tuples, but instead are used as a hint for
whether an IndexOnlyScan needs to visit the underlying relation. The AO
visibility map is used to track which tuples have been deleted from an
AO table, and thus their actual visibility to a table scan. AO relations
do not use a relation fork for a visibility map, and instead track the
visibility of tuples in an auxiliary table.

Co-authored-by: Adam Berlin <aberlin@pivotal.io>