pg_basebackup: WARN when running in TAR mode.
In GPDB when you take a pg_basebackup in tar mode, the segment will not
have the internal.auto.conf file, nor will any tablespaces have the dbid
appended to their symlinks in pg_tblspc.  The backups are still, in
theory, valid, but the tablespace mapping and internal.auto.conf files
will need to be added manually when extracting the backups.

Co-authored-by: Adam Berlin <aberlin@pivotal.io>