Dump persistent tables before and after resync transition
I have run into several instances where a file will not be replicated
during a full segment recovery. I suspect that those files were in a
strange persistent state before the recovery began, but the system did
not leave behind evidence to tell us what that state might be. The most
likely scenario seems to be that AOCO relfiles are stuck in the state
'PersistentFileSysState_BulkLoadCreatePending' after being created, but
it doesn't appear that AOCO relfiles can even go into that state under
normal circumstances.

This commit adds debugging to help determine the root cause of this
issue. It will dump the state of all the persistent objects in an
'interesting' state on a primary doing a full recovery.  The log
messages will be enabled by default, but can be disabled by setting the
debug_filerep_config_print guc to 'false'.