Errors after XLOG_XACT_PREPARE should clean up tablespace directories.
During CREATE TABLESPACE, if an error occurs after the XLOG_XACT_PREPARE
record has been written, there should be no filesystem changes left
on disk for any of the segments. (master, standby, primary, or mirror)

This change adds the tablespace to be deleted to the TwoPhaseHeaderFile
during StartPrepare, so that EndPrepare can store it into the xlog for
XLOG_XACT_ABORT_PREPARED to clean up mirrors, and also perform the
local deletion.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>