Re-introduce explicit lock release for movedb session lock for QD
We had taken out the explicit lock release from the end of the two phase
commit pertaining to a movedb operation, thinking that it would
automatically be released as a part of abort/commit. Session level locks
are only released automatically in the case of an abort as opposed to a
commit. Also, we need the explicit release for the QD since QEs release the
session level lock inside movedb(), even before the associated
transaction is prepared.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>